Crate dice

The dynamic incremental caching computation engine that powers buckv2.

The computation engine will output values corresponding to given Keys, reusing previously computed values when possible. Keys computations are allowed to request other Keys via a ComputationContext.

Example:

mod c {

    /// declaring computations:
    use async_trait::async_trait;
    use derive_more::Display;
    use dice::{Key, InjectedKey, DiceComputations, DiceDataBuilder, DiceData, DiceTransactionUpdater};
    use std::sync::Arc;
    use allocative::Allocative;
use buck2_futures::cancellation::CancellationContext;

    /// A configuration computation that consists of values that are pre-computed outside of DICE
    pub struct InjectConfigs<'compute, 'd>(&'compute mut DiceComputations<'d>);

    impl<'compute, 'd> InjectConfigs<'compute, 'd> {
        pub async fn get_config(&mut self) -> usize {
            self.0.compute(&ConfigKey).await.unwrap()
        }
    }

    #[derive(Clone, Debug, Display, Eq, Hash, PartialEq, Allocative)]
    #[display("{:?}", self)]
    struct ConfigKey;

    #[async_trait]
    impl InjectedKey for ConfigKey {
        type Value = usize;

        fn equality(x: &Self::Value,y: &Self::Value) -> bool {
            x == y
        }
    }

    pub struct MyComputation<'compute, 'd>(pub &'compute mut DiceComputations<'d>);

    impl<'compute, 'd> MyComputation<'compute, 'd> {
        // declaring a computation function
        pub async fn compute_a(&mut self, a: usize, s: String) -> Arc<String> {
            #[derive(Clone, Display, Debug, Eq, Hash, PartialEq, Allocative)]
            #[display("{:?}", self)]
            struct ComputeA(usize, String);

            #[async_trait]
            impl Key for ComputeA {
                type Value = Arc<String>;

                async fn compute(&self, ctx: &mut DiceComputations, _cancellations: &CancellationContext) -> Self::Value {
                    // request for other computations on the self
                    let n = MyComputation(ctx).compute_b(self.0).await;
                    Arc::new(self.1.repeat(n))
                }

                fn equality(x: &Self::Value,y: &Self::Value) -> bool {
                    x == y
                }
            }

            self.0.compute(&ComputeA(a, s)).await.unwrap()
        }

        // second computation function
        pub async fn compute_b(&mut self, a: usize) -> usize {
                self.0.compute(&ComputeB(a)).await.unwrap()
        }
    }

    #[derive(Clone, Display, Debug, Eq, Hash, PartialEq, Allocative)]
    #[display("{:?}", self)]
    struct ComputeB(usize);

    #[async_trait]
    impl Key for ComputeB {
        type Value = usize;

        async fn compute(&self, ctx: &mut DiceComputations, cancellations: &CancellationContext) -> Self::Value {
            self.0 + InjectConfigs(ctx).get_config().await + ctx.global_data().static_data().len()
        }

        fn equality(x: &Self::Value,y: &Self::Value) -> bool {
            x == y
        }
    }

    pub trait SetInjectedConfig {
        fn inject_config(&mut self, i: usize);
    }

    impl SetInjectedConfig for DiceTransactionUpdater {
        /// ways to inject the precomputed values to dice
        fn inject_config(&mut self, i: usize) {
            self.changed_to(vec![(ConfigKey, i)]).unwrap()
        }
    }

    pub trait StaticData {
        fn static_data(&self) -> &String;
    }

    impl StaticData for DiceData {
        fn static_data(&self) -> &String {
            self.get::<String>().unwrap()
        }
    }

    pub trait SetStaticData {
        fn set_static(&mut self, s: String);
    }

    impl SetStaticData for DiceDataBuilder {
        fn set_static(&mut self, s: String) {
            self.set(s);
        }
    }
}

/// how to use computations
use dice::{Dice, DetectCycles};
use std::sync::Arc;
use c::*;

let mut rt = tokio::runtime::Runtime::new().unwrap();
let mut builder = Dice::builder();
builder.set_static("len4".into());
let engine = builder.build(DetectCycles::Disabled);

// inject config
let mut ctx = engine.updater();
ctx.inject_config(0);

let mut ctx = rt.block_on(ctx.commit());

// request the computation from DICE
rt.block_on(async {
    assert_eq!("aaaaaaaa", &*MyComputation(&mut ctx).compute_a(4, "a".into()).await);
});

let mut ctx = engine.updater();
ctx.inject_config(2);

let mut ctx = rt.block_on(ctx.commit());

// request the computation from DICE
rt.block_on(async {
    assert_eq!("aaaaaaaaaa", &*MyComputation(&mut ctx).compute_a(4, "a".into()).await);
});

Modules

introspection

Interfaces for introspection of the DICE graph

testing

Structs

CancellableJoinHandle

CancellationContext

Context available to the function running inside the future to control and manage it’s own cancellation

CancellationHandle

A handle providing the ability to explicitly cancel the associated ExplicitlyCancellableFuture.

Demand

Supplied value can be passed using this object.

Dice

An incremental computation engine that executes arbitrary computations that maps Keys to values.

DiceComputations

The context for computations to register themselves, and request for additional dependencies. The dependencies accessed are tracked for caching via the DiceCtx.

DiceData

DiceDataBuilder

DiceEquality

DiceInvalidationPath

DiceKeyTrackedInvalidationPaths

The invalidation paths for a key. This is accessible from crate::DiceComputations::get_invalidation_paths()

DiceProjectionComputations

Context for projection key computation.

DiceTransaction

The base struct for which all computations start. This is clonable, and dupe, and can be moved to different runtimes to start computations. All computations on this transaction will see only changes at the most-up-to-date version at the time of creation of this transaction.

DiceTransactionUpdater

The struct for which we build transactions. This is where changes are recorded, and committed to DICE, which returns the Transaction where we spawn computations.

DynKey

A type erased Key. Dice APIs that return key references will pass them as DynKey (unless they can be passed as the specific Key type).

GlobalStats

Provides informative stats over all dice instances.

InvalidationPathEntry

A node in the invalidation path.

LinearRecomputeDiceComputations

OpaqueValue

Computed value which is not directly visible to user.

UserComputationData

Includes all user related computation-specific data.

Enums

ActivationData

Describes the kind of activation, and possibly carries data passed by the key’s evaluation.

DetectCycles

DiceEvent

DiceTrackedInvalidationPath

The invalidation path state for a computation node.

InvalidationSourcePriority

Dice tracks up to two invalidation paths for each node, a normal priority and a high priority one. The high priority one considers root invalidated Keys that have a High priority, the normal priority one will consider both Normal and High. The InvalidationSourcePriority is used only for the priority of the Key as a source of invalidation, all nodes participate in both normal and high for propagating invalidations.

WeakFutureError

Traits

ActivationTracker

An ActivationTracker can be used to identify which keys were either reused or computed during a transaction.

DiceEventListener

InjectedKey

Specialized version of Key above. This type of Key is never computed. It should always be injected onto the graph before being requested via DiceCtx::changed_to(Key, Value). Therefore, the compute function on this Key is never called.

Key

The computation Key that maps to a value. The key will be used as an index for caching the computed values.

ProjectionKey

Synchronously computed key from an “opaque” value.

UserCycleDetector

A UserCycleDetector can be used for custom cycle detection in the DICE computation.

UserCycleDetectorGuard

A UserCycleDetectorGuard is used to track the currently computing key. User code can access this through ComputationData::cycle_guard() (and then downcast it with as_any to potentially access custom cycle behavior).