dice

Struct DiceTransaction

pub struct DiceTransaction(/* private fields */);

The base struct for which all computations start. This is clonable, and dupe, and can be moved to different runtimes to start computations. All computations on this transaction will see only changes at the most-up-to-date version at the time of creation of this transaction.

This SHOULD NOT be ever stored by computations, or any results of computations.

Implementations

impl DiceTransaction

pub fn equivalent<E>(&self, other: &E) -> bool
where E: DiceEquivalent,

Returns whether the DiceTransaction is equivalent. Equivalent is defined as whether the two Transactions are based off the same underlying set of key states. That is, all injected keys are the same, and the same compute keys are dirtied, and that any computations that occur between the two transactions can be shared.

pub fn equality_token(&self) -> DiceEquality

pub fn into_updater(self) -> DiceTransactionUpdater

Creates an Updater to record changes to DICE that upon committing, creates a new transaction that keeps the same set of user data. This is equivalent to Dice::updater_with_user_data(data) where the data is taken from the current Transaction.