dice

Trait ActivationTracker

pub trait ActivationTracker:
    Send
    + Sync
    + 'static {
    // Required method
    fn key_activated(
        &self,
        key: &DynKey,
        deps: &mut dyn Iterator<Item = &DynKey>,
        activation_data: ActivationData,
    );
}

An ActivationTracker can be used to identify which keys were either reused or computed during a transaction.

Required Methods

fn key_activated( &self, key: &DynKey, deps: &mut dyn Iterator<Item = &DynKey>, activation_data: ActivationData, )

Receives when a key was activated (computed, or reused). The caller will want to downcast the key and deps to types they care about. The caller also receives whatever the key passed to store_evaluation_data (if any).