dice

Trait Key

pub trait Key:
    Allocative
    + Debug
    + Display
    + Clone
    + Eq
    + Hash
    + Send
    + Sync
    + 'static {
    type Value: Allocative + Dupe + Send + Sync + 'static;

    // Required methods
    fn compute<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        ctx: &'life1 mut DiceComputations<'_>,
        cancellations: &'life2 CancellationContext,
    ) -> Pin<Box<dyn Future<Output = Self::Value> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn equality(x: &Self::Value, y: &Self::Value) -> bool;

    // Provided methods
    fn key_type_name() -> &'static str { ... }
    fn validity(_x: &Self::Value) -> bool { ... }
    fn storage_type() -> StorageType { ... }
    fn provide<'a>(&'a self, demand: &mut Demand<'a>) { ... }
    fn invalidation_source_priority() -> InvalidationSourcePriority { ... }
}

The computation Key that maps to a value. The key will be used as an index for caching the computed values.

Cloning

The Key will need to be cloned. This occurs whenever the key is recomputed and needs to be stored. During normal lookups, only a reference is needed.

The result also needs to be cloned whenever it is requested from DICE. Cloning therefore should be cheap.

Required Associated Types

type Value: Allocative + Dupe + Send + Sync + 'static

Required Methods

fn compute<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, ctx: &'life1 mut DiceComputations<'_>, cancellations: &'life2 CancellationContext, ) -> Pin<Box<dyn Future<Output = Self::Value> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn equality(x: &Self::Value, y: &Self::Value) -> bool

If computed value is equal to previously cached value, DICE won’t invalidate graph nodes depending on this node.

It is safe to return false from this function when values are equal, but returning true when values are not equal would result in inconsistent graph state.

Provided Methods

fn key_type_name() -> &'static str

Provides a short informative name for this key type.

fn validity(_x: &Self::Value) -> bool

If the computed value is false, DICE will consider that result to be a transient value that won’t be re-used on subsequent computations. It will, however, reuse that value for all on-going computations at the current version.

The default here is true, but computations should override this if its expected that they may occasionally produce transient values.

fn storage_type() -> StorageType

fn provide<'a>(&'a self, demand: &mut Demand<'a>)

Machinery to provide. Default implementation is no-op.

Provided value can be obtained with DynKey.

fn invalidation_source_priority() -> InvalidationSourcePriority

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.