dice

Trait ProjectionKey

pub trait ProjectionKey:
    Allocative
    + Clone
    + PartialEq
    + Eq
    + Hash
    + Display
    + Debug
    + Send
    + Sync
    + 'static {
    type DeriveFromKey: Key;
    type Value: Allocative + Dupe + Send + Sync;

    // Required methods
    fn compute(
        &self,
        derive_from: &<<Self as ProjectionKey>::DeriveFromKey as Key>::Value,
        ctx: &DiceProjectionComputations<'_>,
    ) -> Self::Value;
    fn equality(x: &Self::Value, y: &Self::Value) -> bool;

    // Provided methods
    fn validity(x: &Self::Value) -> bool { ... }
    fn storage_type() -> StorageType { ... }
    fn key_type_name() -> &'static str { ... }
}

Synchronously computed key from an “opaque” value.

Required Associated Types

type DeriveFromKey: Key

Key of the value that this projection key is computed from.

type Value: Allocative + Dupe + Send + Sync

This projection key is mapped to this value.

Required Methods

fn compute( &self, derive_from: &<<Self as ProjectionKey>::DeriveFromKey as Key>::Value, ctx: &DiceProjectionComputations<'_>, ) -> Self::Value

Compute the projection key value.

fn equality(x: &Self::Value, y: &Self::Value) -> bool

Provided Methods

fn validity(x: &Self::Value) -> bool

fn storage_type() -> StorageType

fn key_type_name() -> &'static str

Provides a short informative name for this projection type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.