dice

Trait UserCycleDetector

pub trait UserCycleDetector:
    Send
    + Sync
    + 'static {
    // Required methods
    fn start_computing_key(
        &self,
        key: &DynKey,
    ) -> Option<Arc<dyn UserCycleDetectorGuard>>;
    fn finished_computing_key(&self, key: &DynKey);
}

A UserCycleDetector can be used for custom cycle detection in the DICE computation.

Required Methods

fn start_computing_key( &self, key: &DynKey, ) -> Option<Arc<dyn UserCycleDetectorGuard>>

Called by DICE when it starts computing a key. key will be a user Key type (and so user can reliably downcast it to known types).

fn finished_computing_key(&self, key: &DynKey)

Called by DICE when the key finished computing.