dice

Trait UserCycleDetectorGuard

pub trait UserCycleDetectorGuard:
    AsAnyArc
    + Send
    + Sync
    + 'static {
    // Required methods
    fn add_edge(&self, key: &DynKey);
    fn type_name(&self) -> &'static str;
}

A UserCycleDetectorGuard is used to track the currently computing key. User code can access this through ComputationData::cycle_guard() (and then downcast it with as_any to potentially access custom cycle behavior).

Required Methods

fn add_edge(&self, key: &DynKey)

Called by dice when a dependency edge is encountered.

fn type_name(&self) -> &'static str

Used in error messages.